Successfully transmitting spell damage done

+++ Backend/src/modules/live_data_processor/dto/aura_application.rs
@@ -3,6 +3,6 @@ pub struct AuraApplication {
   pub caster: u64,
   pub target: u64,
   pub spell_id: u32,
-  pub stack_amount: u8,
+  pub stack_amount: u32,
   pub applied: bool
 }
